third_party/blink/web_tests/http/tests/navigation/ping-cross-origin-from-https-expected.txt

PingLoader dispatched to 'http://example.test:8000/navigation/resources/save-Ping.php?test=cross-origin-from-https'.
CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/navigation/resources/ping-cross-origin-from-https-target.html' was loaded over HTTPS, but requested an insecure hyperlink auditing endpoint 'http://example.test:8000/navigation/resources/save-Ping.php?test=cross-origin-from-https'. This request has been blocked; the content must be served over HTTPS.
This page calls testRunner.notifyDone().
